Changed logic for storing bridge interface id on MACs

+++ include/NetworkInterface.h
@@ -300,12 +300,12 @@ class NetworkInterface {
   Flow* findFlowByKey(u_int32_t key, AddressTree *allowed_hosts);
   bool findHostsByName(lua_State* vm, AddressTree *allowed_hosts, char *key);
   bool findHostsByMac(lua_State* vm, u_int8_t *mac);
-  bool dissectPacket(u_int8_t bridge_iface_idx,
+  bool dissectPacket(u_int32_t bridge_iface_idx,
 		     u_int8_t *sender_mac, /* Non NULL only for NFQUEUE interfaces */
 		     const struct pcap_pkthdr *h, const u_char *packet,
 		     u_int16_t *ndpiProtocol,
 		     Host **srcHost, Host **dstHost, Flow **flow);
-  bool processPacket(u_int8_t bridge_iface_idx,
+  bool processPacket(u_int32_t bridge_iface_idx,
 		     const struct bpf_timeval *when,
 		     const u_int64_t time,
 		     struct ndpi_ethhdr *eth,
